feature(bank-sdk): Validators refactor
PP-763
