+++ bank-sdk/sdk/src/main/java/net/gini/android/bank/sdk/capture/skonto/validation/SkontoAmountValidator.kt
@@ -1,18 +1,10 @@
-package net.gini.android.bank.sdk.capture.skonto.usecase
+package net.gini.android.bank.sdk.capture.skonto.validation
 
 import net.gini.android.bank.sdk.capture.skonto.SkontoScreenState
 import java.math.BigDecimal
 
-/**
- * Use case for validating the Skonto amount.
- */
-internal class GetSkontoAmountValidationErrorUseCase {
+internal class SkontoAmountValidator {
 
-    /**
-     * Validates the Skonto amount.
-     *
-     * @return [SkontoScreenState.Ready.SkontoAmountValidationError] if the validation fails, null otherwise.
-     */
     fun execute(
         newSkontoAmount: BigDecimal,
         fullAmount: BigDecimal
